Added Database Access Layer Classes
I've included the database classes which I originally developed for
use with Atlantic Lab Equipment's inventory management system. There's a
PDO implementation of the Db class, the QueryPreparer class and its
interface, the PreparedStatement class, the QueryField class, which is
useful for building queries via Db, and the QueryPreparer class, which
is concerned with building the query string based on the state of a Db
instance.

Also included here is the DbFactory class, from which I removed some
unused methods. Its only purpose is to make the creation of a Db
instance easier.