Added the dependencies of the Router class
The Router requires an instance of Logger, UserManager, and Request.
They can be bundled via the dependency container.

The AsyncResponse class serves to keep the JSON interface between client
and server uniform and predictable. Controller methods should be
returning instances of this class if they wish to implement such an
interface.

The exceptions.php file contains user-defined exceptions.