Added Router, Output classes
I've added the Router and Output classes I developed for use with
Atlantic Lab Equipment's inventory management system. I've added two new
methods to the Router class:

Router::init() serves the same purpose as the beginning of ALE IMS's
index.php file: it checks that the user is logged in before creating an
instance of the Router class. This is a static method which returns an
instance of its containing class.

Router::routeAndExecuteRequest() provides a simple method via which to
route requests using the class. Other setups can be configured for use
in lieu of this method via the rest of the Router API.