Added/Renamed Components, Prepared Package for Testing
The structure of the project's filesystem was tweaked to conform to the
PSR-4 autoloading specification (this library is meant to be included in
other projects via Composer).

Added composer.json

Changed the .ini file. The way controllers will be loaded needs to be
tested. There's also an apparent need to include at least one controller
(for authentication purposes) with the package.

The way the Router determines if a request is authorized has changed. A
"Guest" will no longer be logged into a "Guest" account by Sentinel. In
order to access the system via PHP CLI, one must first request to log
in. The mechanics of this need to be tested. I'm not sure if log-ins
work the same via CLI as they do for external clients. Otherwise, it
will be necessary to revert that change and log in the CLI user
automatically when attempts are made to access the system via CLI.

The custom-defined exceptions were moved to their own files to conform
to PSR-4.