Added a JSON Configuration File
I've decided to drop the .ini config file in favor of a JSON file. I
prefer the structure that JSON affords, despite the fact that comments
cannot be written into the file itself. This limitation can be overcome
via separate documentation. Additionally, the configurations will be
easier to access via the Config class, in my opinion.