Refactored Settings Class into Config Class
I've combined the Settings class and the Config class because I found
that they had similar responsibilities. The Settings class was being
used to load the settings from the config file, and the Config class was
being used to configure the application. To me, those responsibilities
aren't different enough to require separate classes.

If I recall correctly, the Config class was requiring the Settings class
as a dependency. To me, it makes more sense to keep the similar
responsibilities together under one class.