Converted Tabs to Spaces in Config.php, Settings.php
So far, I've converted the tabs in each of the files in the /Config
directory to spaces. I've also removed the vertical alignments. These
practices aim to make reading diffs easier. It should be easier to see
the parts that have truly changed, and it should stop me from seeing
whole sections of source highlighted even though only a small change was
made.