RoutingDIContainer now requires UserManager on construction.
Added a new dependency to the RoutingDIContainer: an instance of
UserManager. It's often necessary to have access to the user's details
when routing and processing a request, and including it here in the
dependency injection container serves to provide it to the classes that
need it without adding dependencies to each of those other classes.

Of course, the tabs have been converted to spaces and the vertical
alignment has been removed.