Added new class: User
This class represents a single user account. It should be used to modify
or otherwise operate on individual users. For example, it can be used to
activate/deactivate a user.

Some functionality came from the UserManager class, which was intended
to manage the users as a whole, as well as provide login functionality.

I'd originally used tabs and vertical alignment, but I removed them from
the file in an ongoing effort to update the project's source code styling.