Added new class: Activation
This class represents an Activation record. The class was created to
decouple the third-party user-management library from the rest of the
codebase. For example, when the User class activates a user it can refer
to the activation via this class, rather than the class used by the
library. The idea is to make a future replacement of that library easier
to manage by abstracting its functionality with interfaces for the other
parts of the system to use.

I've also ensured that spaces have been used instead of tabs.