Updated Output class's Public Interface
Output::setResponse now takes an instance of ControllerResponse as an
argument. The ControllerResponse class should be returned from the
controller processing the request and serves to provide a standard
interface to responses. Therefore, the responses can be manipulated in
a predictable way by the Output class.

Tabs have been converted to spaces and the vertical alignment removed.