Added class ControllerResponse.
This class represents a response from the controller which processed the
request. It provides an interface through which other parts of the
system can interact with the response in a predictable fashion. It
provides a flag to denote whether the request was successfully
fulfilled, and it also includes fields for a CLI message, data to
deliver to the client, and a field for plaintext/HTML content for the
template/static page use-case.