Added Interface: IRequestController
Defined an interface for the request-handling controllers.

When the user of this router defines their own controllers they should
conform to this interface in order to ensure their compatibility with
the rest of the system, and to ensure the controllers have each of the
dependencies which may be necessary to process requests.