Updated Dependencies, added PHPUnit as Dev Dependency
Composer could not load some of the dependencies in the composer.json
file. In one instance a newer version of PHP was required. In two other
instances the latest versions did not satisfy my stability requirements.
I've solved these issues for now by requiring older versions of these
dependencies, but I'd like to see if I improve the situation.

I've also added PHPUnit as a development dependency since I'll be using
it to write unit tests.