Changed Visibility of Router::__construct() to Public
In the event that a user would like to construct their own instance of
the Router, perhaps with their own implementations of the dependencies,
they may now do so. One had previously been confined to using the static
init() method.

Moved the __construct() method to the top of the file.

Trailing whitespace has also been removed from the file.