Updated RoutingDIContainer: Removed Db class Dependency
The $db argument to RoutingDIContainer::__construct() is no longer
required to be an instance of class Db. The idea is to allow the user of
the library to use their own database-abstraction layer. Furthermore,
there's a question of whether or not the library itself even needs to
access a database (of course, the user-management library does, but it
takes care of its own DB access, currently via Illuminate/Database).

Also updated the documentation of that method.