Added Dependencies to Router::init(), Removed Dependency on Db class
Router::init() now accepts optional dependencies: an instance of a class
implementing ILogger, and an optional database-abstraction layer
instance (to be supplied to each request-handling controller).

Logger has been updated to implement a yet-to-be-written interface,
ILogger. The database methods have been replaced with an empty method,
Logger::addEntryToDatabase(), which does nothing unless the class is
extended. The aforementioned Logger dependency exists so that the user
of this system can pass their own Logger implementation to the router to
enable database logging. Otherwise, the default Logger implementation
will only log errors/notices to a log file.

Router::init() no longer constructs a Db instance. Database access is
not required by the router, and if a user needs database access in one
of their request-handling controllers, they can pass their db-abstraction
layer to Router::init(). The system will supply that instance to each
controller.

RoutingDIContainer has been updated to reflect the changes above: it
takes an instance of ILogger, rather than Logger, and the $db attribute
is optional, and up to the user to supply an instance of any class they
please.