Removed Global Constants.
The Config class no longer sets global constants. So far, only the
Request class has had need for them, so I added an instance of Config as
a dependency of the Request class. Each reference to a global constant
in Request's methods has been replaced with a call to
Config::getDirective(). The private method Config::setGlobalConstants()
has been removed.

Additionally, I've updated two keys in the config file and updated the
references to them in the Request class.