Moved Permission-check logic to Router::authorizeRequest()
Found permission-checking logic in the method
Router::routeAndExecuteRequest(). Moved it to Router::authorizeRequest.
Now the former method only has to call the latter method once to
determine if the current user should be allowed to continue with the
request.

Some exceptions in Router were also reformatted, and a parameter was
renamed in Router::routeAndExecuteRequest() (serveClientAppOnSync was
renamed to serveClientAppOnSyncReq).

In the UserManager class, the authorize() method was renamed to
hasPermission(). The term "authorize" is being used to describe user
access to the system as a whole, while "permission" refers to individual
request-controller "actions."