Changed UserManager::authorize() to UserManager::authorizeAction()
Also updated a call to that method in the Router class. Some of the
documentation and error messags were also refined.

N.B.:
The call to the above method in the Router class was originally
to UserManager::hasPermission, which was an older name of
UserManager::authorize, and had been resulting in PHP Fatal errors.
Thus, the call has been fixed at the same time as the method was
renamed. There are no other calls to that method at this moment.