Updated the Method Authentication, Permissions Config
To authenticate a user's use of a method, UserManager::authorizeAction()
now accepts the name of the class and the name of the method as separate
arguments.

In the conf.json file, the `permissions` object should contain child
objects labelled with the fully qualified name of the class. Those
objects should contain child arrays labelled by a method name. The arrays
list the permissions required to use that method.

The request controllers are no longer required to have unique
implementations of __toString(). The abstract Controller class now
features a final definition which returns __CLASS__.