Corrected User Activation Methods, Other Fixes
User::getActivation() was updated to make use of the correct Sentinel
activation class. Other methods in that class were updated in the same
fashion.

The Logger class and its interface were corrected - one of the methods
was not requiring the correct Exception class.

The Router CLI methods were updated to log errors which occur during
execution of a CLI request.

The ControllerResponse constructor's validator was throwing exceptions
when the optional arguments were null. This has been corrected.