Output modelled by ControllerResponse, handled by Router
The Router has been updated to deal with output generation via the
ControllerResponse class, which is used to model all output. Now,
all output must be submitted to the Router via Router::setResponse(),
which only accepts instances of that class. Router::serveContent() uses
the Request object to determine how to send the content contained in
that response.

Output generation and emmission is therefore centralized in one place,
and does not need to be implemented for each method which generates
output. Instead, those methods need only return an instance of
ControllerResponse.

It is also possible for the user to define error messages for each
method. Any exception thrown elsewhere in the system, however, will
result in a generic error message.