Updated AdminController::updateUser()
The method will now accept only those data that are to be updated. If
the method cannot find any data to update it will throw an Exception. It
will also allow a change of password, though this is only to ensure it
is possible. If it is indeed possible, that functionality should be
moved to its own method. In addition, that new method should require
"password" and "confirm password" arguments.